move api key atoms to dedicated file
